requirements.txt
jenkinsapi==0.2.16
pyserial==2.7
pytz==2013.8
requests==2.0.1
wsgiref==0.1.2
